New raider dungeon / Oxygen fixes in the sewers (#458)
* New raider dungeon / O2 fixes

* Update Pahrump-Underground-1.dmm

* Update

* Minor changes,

changed the barricades. to metal walls. Removed two drug spawns in the bunker, Fixed a buttons positioning.
(been playing it over and over looking for issues!!

* Minor changes #2

Removed a few bits of railing that can be used to absolutely cheese claws in the sewers

* Minor Fixes

Reconnected the booths in the NCR canteen,

Replaces a ranged raider with a melee raider